Green the CI, and the two things it was hiding
702 passing, ruff clean, the demo spine runs. CI was failing on lint and
stopping there, so two further breakages downstream had never been reported.

The lint itself was mostly cosmetic, but not entirely. Twenty-one duplicate
keys in one test fixture, where the later value silently wins and the earlier
one is dead code nobody can see. Three separate blocks of that dict had been
copied on top of each other by a careless replace of mine. It is rebuilt from
the set of parameters the templates actually declare, and a test now asserts
the fixture covers exactly that set: no duplicates, nothing missing, nothing
kept alive after the template that wanted it was deleted.

One was a real defect in shipping code: `running` in the bill lane, computed
every time a bill arrived and never used, left behind when the running-total
line moved into the settlement block.

Then the two things the lint failure was masking.

demo_spine sent its WhatsApp message to a hardcoded number while registering
the contact from ANBU_DEMO_FAMILY_E164 — my change, a day old. With the
override set they diverge, the send finds no contact, consent refuses it, and
the script falls over reading a key a refusal does not carry. One constant now
feeds both, which is what the divergence argues for.

And assemble_claim_packet gained two required arguments at some point without
the spine following. It passes the admission dates it already narrates.

The rest: int(round(...)) where round already returns an int, sorted()[-1]
where max() says it, unpacked variables nobody reads, a file opened without
closing in a test that runs hundreds of times, and two font loaders catching
bare Exception where OSError is the thing that actually happens. Verified the
fixtures still render at the same byte sizes afterwards, since a font loader
that silently stops working produces blank documents rather than an error.

+++ anbu_care/bills/extract.py
@@ -148,7 +148,7 @@ def _coerce_amount(value: object) -> int | None:
     if isinstance(value, int):
         return value if value >= 0 else None
     if isinstance(value, float):
-        return int(round(value)) if value >= 0 else None
+        return round(value) if value >= 0 else None
     if isinstance(value, str):
         cleaned = re.sub(r"[^\d.]", "", value)
         if not cleaned or cleaned.count(".") > 1:
@@ -157,7 +157,7 @@ def _coerce_amount(value: object) -> int | None:
             number = float(cleaned)
         except ValueError:
             return None
-        return int(round(number)) if number >= 0 else None
+        return round(number) if number >= 0 else None
     return None
